+++ dev/shapes_modifiers.R
@@ -102,6 +102,33 @@ get_slide_number <- function(l)
 }
 
 
+is_com_object <- function(x) {
+  isS4(x) | inherits(x, "COMIDispatch")  
+}
+
+
+stop_if_not_is_com_object <- function(x) {
+  if (!is_com_object(x)) {
+    stop("You must supply a pointer to a COM object.")
+  }
+}
+
+
+get_attribute <- function(x, att, default = NA) 
+{
+  stop_if_not_is_com_object(x)
+  
+  # opts <- options(show.error.messages = FALSE)
+  # on.exit(options(opts))
+  tryCatch({
+    x[[att]]
+  }, error = function(e) {
+    warning(e)
+    default
+  })
+}
+
+
 ppt <- PPT.Open("inst/template.pptx", method = "RDCOMClient")
 l <- get_shapes(ppt)
 shape_info(l)
@@ -110,6 +137,39 @@ visible(l)
 visible(l[2:3]) <- 0
 
 
+get_attribute(s, "sx")
+
+l <- get_shapes(ppt)
+s <- l[[1]]
+# shape attributes
+atts <- c("id", "Name", "ZOrderPosition", "Width", "Height", 
+          "Left", "Top", "Title", "Type", "Visible", 
+          "HasTextFrame", "TextFrame", "TextFrame2", "Fill",
+          "Parent", "Table")
+att_list <- list()
+for (att in atts) {
+  obj <- get_attribute(s, att)
+  att_list[att] <- list(obj)
+} 
+
+
+# draw shapes on side ------------------------------------
 
 
 
+slides <- get_slides(ppt)
+sld <- slides[[4]]
+shapes <- get_shapes_on_slide(sld)
+get_shape_position(shapes[[1]])
+
+# draw slide and all shapes on it
+# get slide
+
+ppt$ppt[["Width"]]
+ppt$ppt[["Height"]]
+ppt$ppt[["ActivePresentation"]][["PageSetup"]][["SlideWidth"]]
+ppt$ppt[["ActivePresentation"]][["PageSetup"]][["SlideHeight"]]
+
+lapply(shapes, z_order)
+s <- shapes[[3]]
+z_order(s) <- 0